% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs_palette}
\alias{obs_palette}
\title{ONSV Color Palette}
\format{
\subsection{\code{obs_palette}}{

A list with 6 colors: blue, yellow, red, green, pink, and orange
}
}
\source{
\url{https://onsv.org.br}
}
\usage{
obs_palette
}
\description{
A list including the hex code of each color used in ONSV style
}
\keyword{datasets}
