% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_discrete_obs.R
\name{scale_discrete_obs}
\alias{scale_discrete_obs}
\title{Apply ONSV color scale to ggplot object}
\usage{
scale_discrete_obs()
}
\value{
A custom scale object for ggplot objects
}
\description{
\code{scale_discrete_obs()} is used to apply the ONSV color scale to a
ggplot graphics
}
\examples{
library(ggplot2)

ggplot(data = iris, aes(x = Sepal.Width, y = Sepal.Length, color = Species)) +
  geom_point() +
  theme_onsv() +
  scale_discrete_obs()
}
