% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_discrete_onsv.R
\name{scale_discrete_onsv}
\alias{scale_discrete_onsv}
\title{Apply ONSV color scale to ggplot object}
\usage{
scale_discrete_onsv(...)
}
\arguments{
\item{...}{Arguments passed on to \link[ggplot2]{scale_discrete_manual} or
to \link[ggplot2]{discrete_scale}, on a lower level.}
}
\value{
A custom scale object for ggplot objects
}
\description{
\code{scale_discrete_onsv()} is used to apply the ONSV color scale to a
ggplot graphics
}
\examples{
library(ggplot2)

ggplot(data = iris, aes(x = Sepal.Width, y = Sepal.Length, color = Species)) +
  geom_point() +
  theme_onsv() +
  scale_discrete_onsv()
}
