% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_obs.R
\name{theme_obs}
\alias{theme_obs}
\title{Apply ONSV theme to ggplot object}
\usage{
theme_obs()
}
\value{
A custom \code{theme()} object.
}
\description{
\code{theme_obs()} is used to apply the ONSV theme to a ggplot graphics.
}
\examples{
library(ggplot2)

ggplot(data = iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  geom_point() +
  theme_obs()
}
