% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_onsv.R
\name{theme_onsv}
\alias{theme_onsv}
\title{Apply ONSV theme to ggplot object}
\usage{
theme_onsv(base_size = 11, ...)
}
\arguments{
\item{base_size}{A number to define font base size,
passed to \link[ggplot2]{theme_minimal}}

\item{...}{Arguments passed on to \link[ggplot2]{theme}}
}
\value{
A custom \code{theme()} object.
}
\description{
\code{theme_onsv()} is used to apply the ONSV theme to a ggplot graphics.
}
\examples{
library(ggplot2)

ggplot(data = iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  geom_point() +
  theme_onsv()
}
